Level-3/requirements.txt
openai>=1.0.0
requests>=2.31.0
litellm>=1.0.0
python-dotenv>=1.0.0
python-dotenv
